Moved route-settings adapter base to separate package
- same approach as the redirects-base extraction (#29279): the store
  contract now lives in @tryghost/adapter-base-route-settings so the
  upcoming GCS store, migration tooling, and custom adapters can depend
  on it without reaching into ghost/core internals
- src/base.ts owns the RouteSettings domain type tree, the
  RouteSettingsStore interface, and an abstract RouteSettingsStoreBase,
  replacing the RouteSettingsStoreBase.js + .d.ts shim pair in core
- routes aren't wired into the adapter-manager yet, so the base is born
  package-based instead of being extracted later — the wiring lands in
  the next commit
- parsing, validation, and serialization stay in ghost/core; only the
  shape moved
