--- ghost/core/core/server/adapters/route-settings/RouteSettingsStoreBase.js
@@ -1,8 +1,0 @@
-module.exports = class RouteSettingsStoreBase {
-    constructor() {
-        Object.defineProperty(this, 'requiredFns', {
-            value: ['get', 'replace'],
-            writable: false
-        });
-    }
-};
