+++ ghost/core/package.json
@@ -87,6 +87,7 @@
     "@slack/webhook": "7.1.0",
     "@tryghost/adapter-base-cache": "0.1.26",
     "@tryghost/adapter-base-redirects": "workspace:*",
+    "@tryghost/adapter-base-route-settings": "workspace:*",
     "@tryghost/adapter-base-sso": "workspace:*",
     "@tryghost/admin-api-schema": "4.7.5",
     "@tryghost/api-framework": "catalog:",
