+++ ghost/core/test/unit/server/services/route-settings/route-settings-fixture.ts
@@ -1,5 +1,6 @@
+import type {RouteSettings} from '@tryghost/adapter-base-route-settings';
+
 import {serializeRouteSettings} from '../../../../../core/server/services/route-settings/route-settings-parser';
-import type {RouteSettings} from '../../../../../core/server/services/route-settings/route-settings-parser';
 
 /**
  * Builds a RouteSettings for tests with a real YAML source derived from the
