+++ packages/adapters/route-settings-base/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./base.ts"
