+++ packages/adapters/route-settings-base/test/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "rootDir": "..",
+        "noEmit": true
+    },
+    "include": [
+        "../src/**/*",
+        "**/*"
+    ]
+}
