+++ packages/adapters/route-settings-base/vitest.config.ts
@@ -1,0 +1,4 @@
+import {createVitestConfig} from '@internal/cfg-vitest';
+
+// Pass overrides to tune coverage thresholds, setupFiles, etc.
+export default createVitestConfig();
